get friend : return user + back not return passwords
