+++ back/src/friends/dtos/FriendDto.ts
@@ -1,6 +1,6 @@
+import { User } from "src/user/user.entity"
+
 export class FriendDto {
-	id: number
-	username: string
-	status: number
+	friend_user: User
 	friend_status: number
 }
