+++ back/src/game/game.gateway.ts
@@ -1,4 +1,3 @@
-import { Logger } from "@nestjs/common";
 import { ConnectedSocket, MessageBody, OnGatewayConnection, OnGatewayDisconnect, SubscribeMessage, WebSocketGateway, WebSocketServer } from "@nestjs/websockets";
 import { Server, Socket } from "socket.io";
 import { UserDto } from "src/user/dtos/UserDto";
