+++ front/package.json
@@ -16,7 +16,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-modal": "^3.16.1",
-    "react-router-dom": "^6.20.0",
+    "react-router-dom": "^6.20.1",
     "react-scripts": "5.0.1",
     "socket.io-client": "^4.7.2",
     "typescript": "^4.9.5",
