+++ front/src/pages/Chat/component/FriendsCardChat.tsx
@@ -6,6 +6,7 @@ import Friend from "../../Friend/interface/friendDto";
 import { useFriends } from "../../Friend/useFriends";
 import { SyntheticEvent } from "react";
 import { useSocket } from "../../../ui/organisms/SocketContext";
+import { Account } from "../../../ui/types";
 
 
 // interface Channel {
@@ -13,7 +14,7 @@ import { useSocket } from "../../../ui/organisms/SocketContext";
 // }
 
 // Utilisation de l'interface dans la fonction FriendCard
-const FriendCardChat: React.FC<{ friend: Friend, set_channel: string }> = ({ friend, set_channel }) => {
+const FriendCardChat: React.FC<{ friend: Account, set_channel: string }> = ({ friend, set_channel }) => {
 	const socket = useSocket();
 	const navigate = useNavigate();
 	const { getStatusColor } = useFriends()
@@ -28,8 +29,8 @@ const FriendCardChat: React.FC<{ friend: Friend, set_channel: string }> = ({ fri
 
 	const createDMChannel = (e: SyntheticEvent) => {
 		e.preventDefault();
-			if (socket) {
-				socket.emit("createDMChannel", friend.id, set_channel);
+		if (socket) {
+			socket.emit("createDMChannel", friend.id, set_channel);
 		}
 	}
 
@@ -40,7 +41,7 @@ const FriendCardChat: React.FC<{ friend: Friend, set_channel: string }> = ({ fri
 					src="https://cdn.intra.42.fr/users/9f5331cff289327a4c7d42c2a66884de/kdi-noce.jpg" />
 			</div>
 			<div className="h-full w-3/5 flex justify-center items-center" onClick={() => handleNav(friend.id)} >
-				<h2>{friend.username.slice(0,8)}</h2>
+				<h2>{friend.username.slice(0, 8)}</h2>
 			</div>
 			<div
 				className="h-5 w-10 rounded-md mr-5 flex justify-center items-center cursor-pointer border-1"
@@ -52,7 +53,7 @@ const FriendCardChat: React.FC<{ friend: Friend, set_channel: string }> = ({ fri
 				onMouseOver={(e) => { e.currentTarget.style.backgroundColor = getStatusColor(friend.status); }}
 				onMouseOut={(e) => { e.currentTarget.style.backgroundColor = ''; }}
 				onClick={createDMChannel}
-				>
+			>
 				<h3 className="text-white">DM</h3>
 			</div>
