+++ front/src/pages/Chat/component/FriendsChat.tsx
@@ -1,34 +1,25 @@
 import React, { useEffect, useState } from 'react';
 import { useSocket } from '../../../ui/organisms/SocketContext';
-import { FriendStatus } from '../../Friend/interface/friendStatus';
-import FriendCard from '../../Friend/FriendCard';
 import { useFriends } from '../../Friend/useFriends';
-import Friend from '../../Friend/interface/friendDto';
 import axios from 'axios';
 import FriendCardChat from './FriendsCardChat';
-
-interface Users {
-    username : string;
-    id : number;
-    friend_status : number;
-    status : number;
-}
+import { Account } from '../../../ui/types';
 
 interface channel {
-    set_channel : string;
+	set_channel: string;
 }
 
-const FriendsChat = ({set_channel} : channel) => {
-	const [users, setUsers] = useState<Users[]>([]);
-	const { getFriends, sortByStatus } = useFriends();
+const FriendsChat = ({ set_channel }: channel) => {
+	const [users, setUsers] = useState<Account[]>([]);
+	const { sortByStatus } = useFriends();
 	const socket = useSocket();
 
 	useEffect(() => {
 		const getUsers = async () => {
 			try {
 				const response = await axios.get("http://localhost:4000/user/users", { withCredentials: true });
-                const sortedUsers = response.data.sort((a : Users, b : Users) => b.status - a.status);
-                setUsers(sortedUsers);
+				const sortedUsers = response.data.sort((a: Account, b: Account) => b.status - a.status);
+				setUsers(sortedUsers);
 			} catch (error) {
 				console.error("Erreur lors de la récupération des users :", error);
 			}
@@ -72,14 +63,14 @@ const FriendsChat = ({set_channel} : channel) => {
 			</div>
 
 			{!users ? (
-				<div style={{ display: "flex", justifyContent: "center", alignItems: "center", height: "70%"}}>
+				<div style={{ display: "flex", justifyContent: "center", alignItems: "center", height: "70%" }}>
 					<h1>No users</h1>
 				</div>
 			) : (
 
 				<div className="h-full m-2.5 bg-black/10 rounded-md	shadow-md shadow-white box-border justify-center items-center overflow-y-auto max-h-[80%]">
-					{users?.map((user: Users) => (
-						<FriendCardChat key={user.id} friend={user} set_channel={set_channel}/>
+					{users?.map((user: Account) => (
+						<FriendCardChat key={user.id} friend={user} set_channel={set_channel} />
 					))}
 				</div>
 			)
