+++ front/src/pages/Friend/component/FriendCard.tsx
@@ -1,11 +1,12 @@
 import { useNavigate } from "react-router-dom";
-import { handleMouseEnter, handleMouseLeave } from "./interface/Tools";
-import Friend from "./interface/friendDto";
-import { useFriends } from "./useFriends";
+import { handleMouseEnter, handleMouseLeave } from "../interface/Tools";
+import Friend from "../interface/friendDto";
+import { useFriends } from "../useFriends";
+import { Account } from "../../../ui/types";
 
 
 // Utilisation de l'interface dans la fonction FriendCard
-const FriendCard: React.FC<{ friend: Friend }> = ({ friend }) => {
+const FriendCard: React.FC<{ friend: Account }> = ({ friend }) => {
 
 	const navigate = useNavigate();
 	const { getStatusColor } = useFriends()
@@ -22,7 +23,7 @@ const FriendCard: React.FC<{ friend: Friend }> = ({ friend }) => {
 		<div className="h-1/5 bg-white/50 m-2.5 rounded-md shadow-lg box-border flex justify-around items-center cursor-pointer" onMouseEnter={handleMouseEnter} onMouseLeave={handleMouseLeave} onClick={() => handleNav(friend.id)}>
 			<div className="h-full w-1/5 flex items-center content-center cursor-pointer">
 				<img alt="image de profil" className="rounded-md h-full"
-					src="https://cdn.intra.42.fr/users/9f5331cff289327a4c7d42c2a66884de/kdi-noce.jpg" />
+					src={friend.avatar} />
 			</div>
 			<div className="h-full w-3/5 flex justify-center items-center" >
 				<h2>{friend.username}</h2>
