+++ front/src/pages/Friend/component/PendingFriendCard.tsx
@@ -1,10 +1,11 @@
 import { Navigate, useNavigate } from "react-router-dom";
-import { handleMouseEnter, handleMouseLeave } from "./interface/Tools";
+import { handleMouseEnter, handleMouseLeave } from "../interface/Tools";
 import { useEffect, useState } from "react";
 import axios from "axios";
-import Friend from "./interface/friendDto";
+import Friend from "../interface/friendDto";
+import { Account } from "../../../ui/types";
 
-const PendingFriendCard = ({ friend }: { friend: Friend }) => {
+const PendingFriendCard = ({ friend }: { friend: Account }) => {
 	const navigate = useNavigate();
 	// const {account} = useAccount();
