+++ front/src/pages/Friend/teststatus.tsx
