+++ front/src/pages/HomePage/HomePage.tsx
@@ -8,7 +8,7 @@ import Ranking from "../Game/gameRanking";
 import { useNavigate } from "react-router-dom";
 import { handleMouseEnter, handleMouseLeave } from "./Tools";
 import AddFriendCard from "./CardContent/AddFriendCard";
-import FriendsToAdd from "../Friend/AddFriend";
+import FriendsToAdd from "../Friend/component/AddFriend";
 
 const HomePage = () => {
